Create python_list_manipulation.py3
In this program, create a list of numbers from 1 to 50 named list_1. The numbers should be present in the increasing order: Ex list_1 = [1,2,3,4,5,....,50]
i.e. index zero should be 1, index one should be 2, index two should be 3 and so on.

Given an input of two numbers, let's say a and b, you have to print the numbers returned by the following command
list_1[a:b]